Create html tags with Emmet
Started with creating the HTML skeleton of the project
by using emmet to avoid the copy paste dillema. Since there is a
need to document step by step process- a slight deviation from the
original process helps me familiarize with the code.

Downloaded the MySQL script file to be executed in preparation for
future database operations.