Recreating Create Process & DB utilities
I want to implement separation of concerns.

Starting with trying to separate the response to an object to
create a standard.

The code for handling the response object involves changing the
values of the response object depending on which
parameters of the CRUD operation is not functioning well.

The DB utilities is relegated to separate class entirely. The same as the
original source.