Remove Dummy content on index & adjust `response`
Adjustments of the setError function to use a string template
and use a parameter with a default value to dynamically change the
error messages and allow for possible custom error messages.

Removed dummy content from index page to allow for tests of content
loading from a mysql table from hr db.