Buggy `get_employees` & orig readability changes
The operation to read the contents of a table and display the contents
with HTML and jquery datatables is working. Although there is a minor
display bug that does not affect the performance of the table.

visually noticeable but functionally safe.

Remove unnecessary php ending tag for visibility and readability improvement.