Change position of EDIT & DELETE | Add DEBUG CODE
Adding code to help with debugging. To execute mysql
and catch error, create a fucntion to execute the query
and will handle the error from the execution of the query.

The function should return the error details and/or the result set
from the query. This makes debugging easier for the delete and edit
operations.

Change the position of the EDIT & DELETE buttons on the LEFT-MOST side
of the table for ease of access with large tables which would push the
content to flow outside the width of the screen.