EDIT Table entry: Sending DATA from SESSION
Trying to send data from the SESSION started from the
php script to generate the data table.

The browser should accept the data and generate the preliminary data
for the update process. Which is to provide the user a placeholder for which
to view the previous content of the field that the user would edit.

The data from which will come from the locally stored data that will be
stored on the browser when the user stores the data.

Modified Event Listener to listen for a class for the Edit and Delete
operation to remove warnings of duplicate id of generated elements.