Separate JS from Index.php
Separating the JavaScript of the Index.php because
I will need to prepare for debugging when creating an
Add_entry feature.

Imports are handled through the index.js, aside from the jquery,
it cannot be imported through other js files. Got stuck but it is
possible, not going to bother with it now, maybe later.