User Log-In process
The original Log-In page implementation used a php file
to handle the log-in process.

Combination of php, javascript, and html to handle the process, I want to
separate the three because I want to use webpack to bundle the whole project
during the final phases of completion.

I could not make proper use of the session to handle user log-in data.

I tried to save the log-in status in session and then use client side
javascript to access the session data, but I could not make
this process work.

So I made the server side php script handle verification then echo out a
a response object containing the status. I tried to handle the redirect
process on the serverside php script but I could not make it work so
the client side javascript will handle the redirection to the
table view page after user log-in.

Started to re-structure the folders and files in preparation for webpack bundling.