Import Fix & Table Generation AJAX comatibility
Making the generate_table_contents.php to only handle
post request from ajax.

Accepting a POST request and then responding with the generated table data.

I chose to use a POST request because it is the recommended methed to be used
over sensitive transaction of data, and it makes it possible to send
complex data for security checks and user verification checks through
the SESSION.

Originally the generate_table_contents
would modify create variables that holds the data from the generated
table which would then be used to echo out the generated table contents.

Which kind of works, but then it creates a lot of unpredictability and
I want the code for table generation to be as independent as much
as possible so that debugging can be easier.

Which is why I chose to separate the functionality of the
code for table generation and the html page that displays the data
which it generated.

Renamed jquery root folder to jquery instead of js because it
is so confusing.