Fix Imports & Table View JS Code Separation
Separate the JavaScript and HTML of Table View.

I got fucked with debugging when both are together.

Restructuring files and folders will mess up the
code that are highly dependent on relative references
to other files.

The import statements did not work at all because of the
restructure and using the base attribute the code
even more confusing.

The import statements will likely break if the
environment changes. The relative references are
kind of in a dangerous state.