Created a dummy html file for Table View
The dummy table view page is to test out how the
actual table view page will be presented
in the browser.

Similar use case to a prototype during development.

The dummy html file is for development purposes only
and is not intended to be presented to the user.

The contents are static and not generated.