Working Display of Different Tables
Though there is a little problem with the generated
table in that there is no content divided available.

For every different load of table content the table
content divider does not reset or present any divider.

The default table to be viewed is the employees
table because I set it as the default table to
be viewed.

The problem is inside the code for table generation.
After commenting out the code that sets the
default table name to employees, it is possible to
change the table to be viewed in the page.