WOrking Display of Selection Input For Add Entry
Display of the selection entry for different tables is working.

For every different table, the add entry modal will also change and
present the appopriate input elements to be used for the add entry process.

The code is still buggy and will cause problems later on development.

The data of the table that contains which columns uses a foreign key
and what is the associated mysql query to generate the options for values
of that column is generated on a separate php file for ease of debugging.

I did not mix it with other files because it causes a lot of problems
and makes it harder to debug.

The process to generate the input element starts with checking and verifying
if the table has columns which uses foreign key constraints, the
columns that rely on other tables for values.

If the table has no column that has a foreign key constraint then it will
be generated through default, generation of input and label elements.