Working Retrieving data from add entry modal
The data from the add entry modal that contains the
input elements for adding an entry to a table can be
retrieved through the save button.

The data will be transformed to json format when it is sent
to the php script that handles the add entry to tabel process.

The add_employee.php will be used for adding an entry to a table.

The original purpose was to add an entry to the employee table,
but it is possible to have it's functionality to add an entry to
other tables aside from the employees table.