Jtable Fix & Add BUGGY add entry process
The add entry process works for all the tables except the
jobs table. I don't fucking know what is wrong, the insert query won't
work with the table but other tables work fine.

Like why???????

The jtable not displaying after the display of the table content is
fixed. During the page load before selecting a table to view, the
`DataTable()` function call shoulld not be executed, the call should
be executed only after the contents of the table is loaded.

The add_employee.php is changed to add_entry.php, instead of adding
an employee entry, it is possible to add an entry for every table aside
from the jobs table. I don't know why it does not work.