Illegal fix to the bug in inserting into jobs
Inserting values into the jobs table causes
an error that the totJobs table does not exists.

So to fix the error and to insert values into
totJobs, then I am adding the totJobs table.

I could just remove the trigger that uses the
totJobs table, but I don't want to delete
and/or remove anything from the database, so
I just added a table to the database.