Create php script for Update Process
For every table that updates an entry, the
php file that will be called should be this file.

This will likely break. to be debugged.

The connection used by the update is a connection
from the session.

I want to use the session to
store a single connection that executes mysql queries
rather than create a new mysql connection
for every query.