Modify TableView page and script for UPDATE ENTRY
I think this is a messy set-up. The event triggers
that uses the id and classes of elements, is fucked.

Originally the id to identify the element for the
trigger for a click event is changed to a class.

I am confused, which of which, but this will
be changed during development because I don't
know which of which is during development.

The TableView page should be split into different pages
for every table.