WORKING UPDATE ENTRY
The update process is working for all tables.

Code is a mess.

I just copied the code for the add entry process for the
update process. Which caused a lot of problems.

There are some changes to make the edit process variables
 be somewhat identifiable from the add process variables.

This caused a lot of problems because I did just copy the
code directly and then after not properly reading the code
and fixing the references, the code blew up hahahahahahhahahaha

This is to be debugged and edited later so the code is not meant for
long term but for the sake of finishing things as fast as possible.