WORKING DELETE ENTRY
I need to sleep. I kind of copied the whole
code from the update process but I removed most
of the code to check and process data to be
sent to the server.

I just sent the table name and identifier-value pair to
the server to delete the corresponding row of the table
inside the server.

Weellll I think this is the fastest portion of the
CRUD that I manage to finish, but the ones with
the bugs that I have not noticed, so will
likely break soon.