MODIFIED Add_entry for USE outside TableView.php
The Add_entry.php has a function that handles the
code to input the data for an entry into the database.

The function `add_entry()` originally uses the $_POST[] array to
access the necessary data, but this makes it so that the name of the
data must be precise otherwise it won't be referenced and will
reflect as an error.

The entry data for input to the database is changed,instead of using
the $_POST array, a parameter is used to channel the data from the
$_POST array.

I have to use the code to add an entry to the database, aside from the
register user page, other pages may need the code.

Direct access to the $_POST array is dangerous, because variable names
must be precise and specific. This can be alleviated through the
fucntion parameters.