Added Utility func for CRUD func in TableView.js
The javascript function for add and update operation in TableView.js
is changed to refer to utility functions to retrieve and process data
to be sent to the server.

Originally the code for adding an entry is copied to the update function,
since I found it a pain to refactor the reusable code portion into a
separate function.

I plan to make some of the crud functions into different javascript files,
so that I can reuse some of the CRUD code. The add_entry function can be
reused by the register page, instead of redefining the function,
I should have exported the function into an external file and imported
the function.