Working Register Page : Base-No-Design
No Design with Bootstrap.

The register page works, can register user to the database.

No response upon successful user registration.

Uses the same code as the Add_entry function in TableView.js,
I recreated the function, I just realized I should have imported
the function.

The rendering of the contents of the option elements is done inside the
script, instead of having the task be processed in the server.

The only change I did is having the data for creating the options is
passed to the client side from the server side.